refactor (#427)
Prepare for more tests
